Mon, Jul 10, 2023, 6:01 PM -04:00

+++ cs7/week3/pset/pset3e/divisors.py
@@ -57,7 +57,7 @@ def divisors(num):
     sum = 0
 
     # If num is divisible by a number <= its square root,
-    for potential_div in range(1, sqrt + 1):
+    for potential_div in range(1, int(sqrt) + 1):
         if num % potential_div == 0:
             sum += potential_div
